refactor: ♻️ Implement major architectural and quality improvements
This commit applies a series of refactors based on our code review, focusing on architecture, testability, and code quality.

Architecture & DRY:
- Moves DiagnosticProgressModel out of the domain layer and into the data layer, cleaning the separation of concerns.
- Fixes model duplication (DRY violation) by importing sub-models (DeviceInfoModel, NetworkInfoModel, SpeedTestResultModel) in FinalResultsModel instead of redefining them.

DI & State Management:
- Standardizes on GetIt for dependency injection.
- Removes Provider for AppConfig and updates all consuming widgets (HomeScreen, DiagnosticButton, QuickTipsCard) to use sl<AppConfig>().

Lifecycle & Resource Management:
- Centralizes resource disposal logic in DiagnosticCubit.
- The Cubit now owns the repository lifecycle and calls disposeResources() on close, onDone, and onError.
- DiagnosticRepositoryImpl no longer disposes the data source.

Testing:
- Fixes the DiagnosticCubit test suite by initializing the DI container (di.init) with useMockDiagnostic = true.
- Ensures SharedPreferences is mocked to prevent MissingPluginException.

Cleanup (Dead Code):
- Removes unused files: diagnostic_screen_old.dart, the empty esults_screen.dart, and the duplicated core/di/error/failures.dart.
