--- lib/core/di/error/failures.dart
@@ -1,15 +1,0 @@
-import 'package:equatable/equatable.dart';
-
-abstract class Failure extends Equatable {
-  const Failure([List properties = const <dynamic>[]]);
-}
-
-class ServerFailure extends Failure {
-  @override
-  List<Object?> get props => [];
-}
-
-class CacheFailure extends Failure {
-  @override
-  List<Object?> get props => [];
-}
