+++ lib/data/models/network_info_model.dart
@@ -12,10 +12,29 @@ class NetworkInfoModel extends NetworkInfoEntity {
     super.internalIP,
   });
 
+  factory NetworkInfoModel.fromEntity(NetworkInfoEntity entity) {
+    return NetworkInfoModel(
+      connectionType: entity.connectionType,
+      wifiName: entity.wifiName,
+      wifiFrequency: entity.wifiFrequency,
+      wifiSignalStrength: entity.wifiSignalStrength,
+      wifiLinkSpeed: entity.wifiLinkSpeed,
+      wifiBSSID: entity.wifiBSSID,
+      externalIP: entity.externalIP,
+      internalIP: entity.internalIP,
+    );
+  }
+
   factory NetworkInfoModel.fromJson(Map<String, dynamic> json) {
     return NetworkInfoModel(
-      connectionType: json['connectionType'],
-      wifiName: json['wifiName'],
+      connectionType: json['connectionType'] as String,
+      wifiName: json['wifiName'] as String?,
+      wifiFrequency: json['wifiFrequency'] as String?,
+      wifiSignalStrength: json['wifiSignalStrength'] as int?,
+      wifiLinkSpeed: json['wifiLinkSpeed'] as int?,
+      wifiBSSID: json['wifiBSSID'] as String?,
+      externalIP: json['externalIP'] as String?,
+      internalIP: json['internalIP'] as String?,
     );
   }
 
@@ -23,6 +42,12 @@ class NetworkInfoModel extends NetworkInfoEntity {
     return {
       'connectionType': connectionType,
       'wifiName': wifiName,
+      'wifiFrequency': wifiFrequency,
+      'wifiSignalStrength': wifiSignalStrength,
+      'wifiLinkSpeed': wifiLinkSpeed,
+      'wifiBSSID': wifiBSSID,
+      'externalIP': externalIP,
+      'internalIP': internalIP,
     };
   }
 }
