+++ lib/domain/entities/diagnostic_flow.dart
@@ -1,6 +1,5 @@
 import 'package:equatable/equatable.dart';
 import 'final_results_entity.dart';
-import '../../data/models/speed_test_result_model.dart';
 
 enum DiagnosticStage {
   initializing,
@@ -78,39 +77,4 @@ class DiagnosticCompleted extends DiagnosticFlowEvent {
 
   @override
   List<Object?> get props => [results];
-}
-
-class DiagnosticProgressModel extends DiagnosticProgressEntity {
-  const DiagnosticProgressModel({
-    required super.stage,
-    required super.progress,
-    required super.message,
-    required super.timestamp,
-    super.speedTestResult,
-  });
-
-  factory DiagnosticProgressModel.fromJson(Map<String, dynamic> json) {
-    return DiagnosticProgressModel(
-      stage: DiagnosticStage.values.firstWhere(
-            (e) => e.toString() == json['stage'],
-        orElse: () => DiagnosticStage.initializing,
-      ),
-      progress: (json['progress'] as num?)?.toDouble() ?? 0.0,
-      message: json['message'] as String? ?? '',
-      timestamp: DateTime.parse(json['timestamp']),
-      speedTestResult: json['speedTestResult'] != null
-          ? SpeedTestResultModel.fromJson(json['speedTestResult'])
-          : null,
-    );
-  }
-
-  Map<String, dynamic> toJson() {
-    return {
-      'stage': stage.toString(),
-      'progress': progress,
-      'message': message,
-      'timestamp': timestamp.toIso8601String(),
-      'speedTestResult': speedTestResult != null ? (speedTestResult as SpeedTestResultModel).toJson() : null,
-    };
-  }
 }
