+++ lib/main.dart
@@ -4,8 +4,6 @@ import 'package:maxt_diagnostic/app/navigation/app_router.dart';
 import 'package:maxt_diagnostic/features/home/presentation/cubit/home_cubit.dart';
 import 'core/di/injection_container.dart' as di;
 import 'l10n/app_localizations.dart';
-import 'package:provider/provider.dart';
-import 'core/config/app_config.dart';
 import 'core/theme/app_theme.dart';
 
 void main() async {
@@ -22,12 +20,8 @@ class MyApp extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    final config = di.sl<AppConfig>();
-    return MultiProvider(
-      providers: [
-        Provider<AppConfig>.value(value: config),
-        BlocProvider(create: (_) => di.sl<HomeCubit>()),
-      ],
+    return BlocProvider(
+      create: (_) => di.sl<HomeCubit>(),
       child: MaterialApp.router(
         title: 'MaxT Diagnostic',
         theme: AppTheme.lightTheme,
