+++ pubspec.yaml
@@ -39,7 +39,6 @@ dependencies:
   # PDF & Sharing
   printing: ^5.12.0
   share_plus: ^7.2.2
-  provider: ^6.1.2
   flutter_dotenv: ^5.1.0
 
   # Storage
